+++ conf/app.routes
@@ -17,7 +17,7 @@ GET         /returns-submission/:id/:periodKey      uk.gov.hmrc.plasticpackaging
 
 # DES obligation data
 GET         /obligation-data/:pptReference          uk.gov.hmrc.plasticpackagingtaxreturns.controllers.ObligationDataController.get(pptReference: String, fromDate: LocalDate, toDate: LocalDate, status: Option[String])
-GET         /obligations/open/:pptReference         uk.gov.hmrc.plasticpackagingtaxreturns.controllers.TileInfoController.get(pptReference)
+GET         /obligations/open/:pptReference         uk.gov.hmrc.plasticpackagingtaxreturns.controllers.PPTObligationController.get(pptReference)
 
 
 # DES financial data
