+++ it/support/AuthTestSupport.scala
@@ -14,6 +14,8 @@
  * limitations under the License.
  */
 
+package support
+
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.when
 import org.mockito.stubbing.OngoingStubbing
