+++ test/uk/gov/hmrc/plasticpackagingtaxreturns/controllers/controllers/PPTObligationControllerSpec.scala
@@ -21,75 +21,72 @@ import org.mockito.Mockito.{reset, verify, when}
 import org.scalatest.BeforeAndAfterEach
 import org.scalatestplus.mockito.MockitoSugar
 import org.scalatestplus.play.PlaySpec
-import org.scalatestplus.play.guice.GuiceOneAppPerSuite
-import play.api.Application
-import play.api.inject.bind
-import play.api.inject.guice.GuiceApplicationBuilder
 import play.api.libs.json.Json
 import play.api.mvc.Result
-import play.api.test.Helpers.{OK, contentAsJson, defaultAwaitTimeout, route, status, _}
-import play.api.test.{FakeRequest, Injecting}
-import uk.gov.hmrc.auth.core.AuthConnector
+import play.api.test.Helpers.{OK, contentAsJson, defaultAwaitTimeout, status, _}
+import play.api.test.{FakeRequest, Helpers}
 import uk.gov.hmrc.plasticpackagingtaxreturns.connectors.ObligationDataConnector
 import uk.gov.hmrc.plasticpackagingtaxreturns.connectors.models.des.enterprise.{Identification, Obligation, ObligationDataResponse, ObligationStatus}
-import uk.gov.hmrc.plasticpackagingtaxreturns.controllers.TileInfoController
-import uk.gov.hmrc.plasticpackagingtaxreturns.controllers.base.AuthTestSupport
-import uk.gov.hmrc.plasticpackagingtaxreturns.controllers.models.SubscriptionTestData
+import uk.gov.hmrc.plasticpackagingtaxreturns.controllers.PPTObligationController
+import uk.gov.hmrc.plasticpackagingtaxreturns.controllers.base.it.FakeAuthenticator
 import uk.gov.hmrc.plasticpackagingtaxreturns.models.PPTObligations
 import uk.gov.hmrc.plasticpackagingtaxreturns.repositories.TaxReturnRepository
 import uk.gov.hmrc.plasticpackagingtaxreturns.services.PPTObligationsService
 
 import java.time.LocalDate
+import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
 
-class TileInfoControllerSpec
-    extends PlaySpec with AuthTestSupport with SubscriptionTestData with BeforeAndAfterEach with MockitoSugar
-    with GuiceOneAppPerSuite with Injecting {
+class PPTObligationControllerSpec
+  extends PlaySpec
+    with BeforeAndAfterEach
+    with MockitoSugar {
 
   val mockPPTObligationsService: PPTObligationsService = mock[PPTObligationsService]
   val mockObligationDataConnector               = mock[ObligationDataConnector]
   val mockTaxReturnRepository                   = mock[TaxReturnRepository]
-  lazy val sut: TileInfoController              = inject[TileInfoController]
 
   override def beforeEach(): Unit = {
     super.beforeEach()
-    reset(mockAuthConnector, mockPPTObligationsService, mockObligationDataConnector, mockTaxReturnRepository)
-
+    reset(
+      mockPPTObligationsService,
+      mockObligationDataConnector,
+      mockTaxReturnRepository)
   }
 
-  override lazy val app: Application = GuiceApplicationBuilder()
-    .overrides(bind[PPTObligationsService].toInstance(mockPPTObligationsService),
-               bind[ObligationDataConnector].toInstance(mockObligationDataConnector),
-               bind[AuthConnector].toInstance(mockAuthConnector),
-               bind[TaxReturnRepository].toInstance(mockTaxReturnRepository)
-    )
-    .build()
+  val cc = Helpers.stubControllerComponents()
+
+  lazy val sut = new PPTObligationController(
+    cc,
+    new FakeAuthenticator(cc),
+    mockObligationDataConnector,
+    mockPPTObligationsService)
 
   "get" must {
-    val request          = FakeRequest("GET", "/obligations/open/" + pptReference)
     val obligations      = PPTObligations(None, None, 0, false, false)
     val rightObligations = Right(obligations)
+    val pptReference = "1234"
 
     "be accessible from the requestHandler" in { //todo eventually be moved to integration test package
-      withAuthorizedUser()
+
       when(mockPPTObligationsService.constructPPTObligations(any())).thenReturn(rightObligations)
       when(mockObligationDataConnector.get(any(), any(), any(), any())(any()))
         .thenReturn(Future.successful(Right(ObligationDataResponse(Seq.empty))))
 
-      val result: Future[Result] = route(app, request).get
+      val result: Future[Result] = sut.get(pptReference).apply(FakeRequest())
 
       status(result) must not be NOT_FOUND
     }
 
     "get PTPObligation from service" in {
-      withAuthorizedUser()
+
       val desResponse = ObligationDataResponse(Seq.empty)
 
       when(mockPPTObligationsService.constructPPTObligations(any())).thenReturn(rightObligations)
       when(mockObligationDataConnector.get(any(), any(), any(), any())(any()))
         .thenReturn(Future.successful(Right(desResponse)))
 
-      val result: Future[Result] = sut.get(pptReference)(request)
+      val result: Future[Result] = sut.get(pptReference).apply(FakeRequest())
 
       status(result) mustBe OK
       verify(mockPPTObligationsService).constructPPTObligations(ObligationDataResponse(Seq.empty))
@@ -97,56 +94,54 @@ class TileInfoControllerSpec
     }
 
     "get should call Obligation Connector" in {
-      withAuthorizedUser()
+
       val desResponse = ObligationDataResponse(Seq.empty)
 
       when(mockObligationDataConnector.get(any(), any(), any(), any())(any()))
         .thenReturn(Future.successful(Right(desResponse)))
       when(mockPPTObligationsService.constructPPTObligations(any())).thenReturn(rightObligations)
 
-      await(sut.get(pptReference)(request))
+      sut.get(pptReference).apply(FakeRequest())
 
       verify(mockObligationDataConnector)
         .get(exactlyEq(pptReference),
-             exactlyEq(LocalDate.of(2022, 4, 1)),
-             exactlyEq(LocalDate.now()),
-             exactlyEq(ObligationStatus.OPEN)
+          exactlyEq(LocalDate.of(2022, 4, 1)),
+          exactlyEq(LocalDate.now()),
+          exactlyEq(ObligationStatus.OPEN)
         )(any())
     }
 
     "get should call the service" in {
-      withAuthorizedUser()
-
       val desResponse = ObligationDataResponse(Seq(Obligation(Identification("", "", ""), Nil)))
 
       when(mockObligationDataConnector.get(any(), any(), any(), any())(any()))
         .thenReturn(Future.successful(Right(desResponse)))
       when(mockPPTObligationsService.constructPPTObligations(any())).thenReturn(rightObligations)
 
-      await(sut.get(pptReference)(request))
+      await(sut.get(pptReference).apply(FakeRequest()))
 
       verify(mockPPTObligationsService).constructPPTObligations(desResponse)
     }
 
     "return internal server error response" when {
       "if error return from connector" in {
-        withAuthorizedUser()
+
         when(mockObligationDataConnector.get(any(), any(), any(), any())(any()))
           .thenReturn(Future.successful(Left(500)))
 
-        val result: Future[Result] = sut.get(pptReference)(request)
+        val result: Future[Result] = sut.get(pptReference).apply(FakeRequest())
 
         status(result) mustBe INTERNAL_SERVER_ERROR
       }
 
       "if the controller handles an error" in {
-        withAuthorizedUser()
+
         val desResponse = ObligationDataResponse(Seq(Obligation(Identification("", "", ""), Nil)))
         when(mockObligationDataConnector.get(any(), any(), any(), any())(any()))
           .thenReturn(Future.successful(Right(desResponse)))
         when(mockPPTObligationsService.constructPPTObligations(any())).thenReturn(Left("test error message"))
 
-        val result: Future[Result] = sut.get(pptReference)(request)
+        val result: Future[Result] = sut.get(pptReference).apply(FakeRequest())
 
         status(result) mustBe INTERNAL_SERVER_ERROR
       }
